% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreRython.R
\name{translate_python_to_r}
\alias{translate_python_to_r}
\title{Translate Python code to R code}
\usage{
translate_python_to_r(python_code)
}
\arguments{
\item{python_code}{A character string of Python code.}
}
\value{
A character string of translated R code.
}
\description{
Translate Python code to R code
}
